shl: do not generate atom tests
It is not working yet, we must implement a method that transforms
the path from a transducer to a trace that we can dump to the test.
This is not that hard, but also not straightforward as we basically
have to execute the path (remember the values of registers and so on).

Executing the path is actually very close to doing symbolic execution,
so we might, in the future, implement that and use it for generating tests.